behaviortree.cpp: fix unvendored dependencies

+++ recipes/behaviortree.cpp/all/conanfile.py
@@ -1,7 +1,6 @@
 from conan import ConanFile
 from conan.errors import ConanInvalidConfiguration
 from conan.tools.env import VirtualBuildEnv
-from conan.tools.microsoft import check_min_vs, is_msvc
 from conan.tools.files import get, copy, rmdir, replace_in_file, save
 from conan.tools.build import check_min_cppstd
 from conan.tools.scm import Version
@@ -40,18 +39,25 @@ class BehaviorTreeCPPConan(ConanFile):
 
     @property
     def _minimum_compilers_version(self):
-        if Version(self.version) < "4.0":
-            return {
-                "gcc": "5",
-                "clang": "5",
-                "apple-clang": "12",
-            }
-        else:
+        if Version(self.version) >= "4.0":
             return {
                 "gcc": "8",
                 "clang": "7",
                 "apple-clang": "12",
+                "msvc": "192",
+                "Visual Studio": "16",
             }
+        else:
+            return {
+                "gcc": "5",
+                "clang": "5",
+                "apple-clang": "12",
+                "msvc": "191",
+                "Visual Studio": "15",
+            }
+
+    def export_sources(self):
+        copy(self, "conan_deps.cmake", self.recipe_folder, os.path.join(self.export_sources_folder, "src"))
 
     def config_options(self):
         if self.settings.os == "Windows":
@@ -64,31 +70,55 @@ class BehaviorTreeCPPConan(ConanFile):
     def layout(self):
         cmake_layout(self, src_folder="src")
 
+    @property
+    def _with_tinyxml2(self):
+        return Version(self.version) >= "3.8.0"
+
+    @property
+    def _with_lexy(self):
+        return Version(self.version) >= "4.0.0"
+
+    @property
+    def _with_sqlite3(self):
+        return Version(self.version) >= "4.1.1"
+
+    @property
+    def _with_minitrace(self):
+        return Version(self.version) >= "4.3.4"
+
     def requirements(self):
         if self.options.with_coroutines:
             self.requires("boost/1.83.0")
         self.requires("ncurses/6.4")
         self.requires("zeromq/4.3.5")
-        self.requires("cppzmq/4.10.0")
-        self.requires("minitrace/cci.20230905")
-        self.requires("tinyxml2/10.0.0")
-        self.requires("foonathan-lexy/2022.12.1")
-        if Version(self.version) >= "4.1.1":
+        if self._with_lexy:
+            self.requires("foonathan-lexy/2022.12.1")
+        if self._with_minitrace:
+            self.requires("minitrace/cci.20230905")
+        if self._with_sqlite3:
             self.requires("sqlite3/3.44.2")
+        if self._with_tinyxml2:
+            self.requires("tinyxml2/10.0.0", transitive_libs=Version(self.version) < "4.0")
+
+        # TODO: other vendored dependencies
+        # - cppzmq is customized and not compatible with Conan version
+        # - cpp-sqlite
+        # - minicoro
+        # - wildcards
 
     def validate(self):
         if self.info.settings.os == "Windows" and self.info.options.shared:
             raise ConanInvalidConfiguration(f"{self.ref} can not be built as shared on Windows.")
         if self.info.settings.compiler.get_safe("cppstd"):
             check_min_cppstd(self, self._minimum_cppstd_required)
-        check_min_vs(self, 191 if Version(self.version) < "4.0" else 192)
-        if not is_msvc(self):
-            minimum_version = self._minimum_compilers_version.get(str(self.info.settings.compiler), False)
-            if not minimum_version:
-                self.output.warn(f"{self.ref} requires C++{self._minimum_cppstd_required}. Your compiler is unknown. Assuming it supports C++{self._minimum_cppstd_required}.")
-            elif Version(self.info.settings.compiler.version) < minimum_version:
-                raise ConanInvalidConfiguration("BehaviorTree.CPP requires C++{}, which your compiler does not support."
-                                                .format(self._minimum_cppstd_required))
+        minimum_version = self._minimum_compilers_version.get(str(self.info.settings.compiler), False)
+        if not minimum_version:
+            self.output.warn(f"{self.ref} requires C++{self._minimum_cppstd_required}. "
+                             f"Your compiler is unknown. Assuming it supports C++{self._minimum_cppstd_required}.")
+        elif Version(self.info.settings.compiler.version) < minimum_version:
+            raise ConanInvalidConfiguration(
+                f"BehaviorTree.CPP requires C++{self._minimum_cppstd_required}, which your compiler does not support."
+            )
 
         if self.settings.compiler == "clang" and str(self.settings .compiler.libcxx) == "libstdc++":
             raise ConanInvalidConfiguration(f"{self.ref} needs recent libstdc++ with charconv. please switch to gcc, or to libc++")
@@ -105,6 +135,10 @@ class BehaviorTreeCPPConan(ConanFile):
         venv.generate()
 
         tc = CMakeToolchain(self)
+        tc.variables["CMAKE_PROJECT_behaviortree_cpp_INCLUDE"] = "conan_deps.cmake"
+        tc.variables["WITH_LEXY"] = self._with_lexy
+        tc.variables["WITH_MINITRACE"] = self._with_minitrace
+        tc.variables["WITH_TINYXML2"] = self._with_tinyxml2
         if Version(self.version) < "4.0":
             tc.variables["BUILD_EXAMPLES"] = False
             tc.variables["BUILD_UNIT_TESTS"] = False
@@ -129,31 +163,23 @@ class BehaviorTreeCPPConan(ConanFile):
         replace_in_file(self, cmakelists, "set(CMAKE_POSITION_INDEPENDENT_CODE ON)\n", "")
         # Disable -Werror
         if Version(self.version) < "4.3":
-            replace_in_file(self, cmakelists, " /WX", "", strict=False)
-            replace_in_file(self, cmakelists, " -Werror=return-type", "", strict=False)
-        # Unvendor minitrace
-        rmdir(self, os.path.join(self.source_folder, "3rdparty", "minitrace"))
-        replace_in_file(self, cmakelists, "3rdparty/minitrace/minitrace.cpp", "")
-        save(self, cmakelists,
-             "\nfind_package(minitrace REQUIRED CONFIG)\n"
-             "target_link_libraries(${BTCPP_LIBRARY} PRIVATE minitrace::minitrace)\n",
-             append=True)
-        replace_in_file(self, os.path.join(self.source_folder, "src", "loggers", "bt_minitrace_logger.cpp"),
-                        "minitrace/minitrace.h", "minitrace.h")
-        # Unvendor tinyxml2
-        rmdir(self, os.path.join(self.source_folder, "3rdparty", "tinyxml2"))
-        replace_in_file(self, cmakelists, "3rdparty/tinyxml2/tinyxml2.cpp", "")
-        save(self, cmakelists,
-             "\nfind_package(tinyxml2 REQUIRED CONFIG)\n"
-             "target_link_libraries(${BTCPP_LIBRARY} PRIVATE tinyxml2::tinyxml2)\n",
-             append=True)
-        replace_in_file(self, os.path.join(self.source_folder, "src", "xml_parsing.cpp"),
-                        "tinyxml2/tinyxml2.h", "tinyxml2.h")
+            replace_in_file(self, cmakelists, " -Werror=return-type", "")
         # Unvendor lexy
-        rmdir(self, os.path.join(self.source_folder, "3rdparty", "lexy"))
-        replace_in_file(self, cmakelists,
-                        "add_subdirectory(3rdparty/lexy)",
-                        "find_package(lexy REQUIRED CONFIG)")
+        if self._with_lexy:
+            rmdir(self, os.path.join(self.source_folder, "3rdparty", "lexy"))
+            save(self, os.path.join(self.source_folder, "3rdparty", "lexy", "CMakeLists.txt"), "")
+        # Unvendor minitrace
+        if self._with_minitrace:
+            rmdir(self, os.path.join(self.source_folder, "3rdparty", "minitrace"))
+            replace_in_file(self, cmakelists, "3rdparty/minitrace/minitrace.cpp", "")
+            replace_in_file(self, os.path.join(self.source_folder, "src", "loggers", "bt_minitrace_logger.cpp"),
+                            "minitrace/minitrace.h", "minitrace.h")
+        # Unvendor tinyxml2
+        if self._with_tinyxml2:
+            rmdir(self, os.path.join(self.source_folder, "3rdparty", "tinyxml2"))
+            replace_in_file(self, cmakelists, "3rdparty/tinyxml2/tinyxml2.cpp", "")
+            replace_in_file(self, os.path.join(self.source_folder, "src", "xml_parsing.cpp"),
+                            "tinyxml2/tinyxml2.h", "tinyxml2.h")
 
     def build(self):
         self._patch_sources()
@@ -170,29 +196,34 @@ class BehaviorTreeCPPConan(ConanFile):
         rmdir(self, os.path.join(self.package_folder, "lib", "cmake"))
 
     def package_info(self):
-        if Version(self.version) < "4.0":
-            self.cpp_info.set_property("cmake_file_name", "BehaviorTreeV3")
+        if Version(self.version) >= "4.0":
+            cmake_file_name = "BehaviorTree"
+            libname = "behaviortree_cpp"
         else:
-            self.cpp_info.set_property("cmake_file_name", "BehaviorTree")
-
-        libname = "behaviortree_cpp_v3" if Version(self.version) < "4.0" else "behaviortree_cpp"
+            cmake_file_name = "BehaviorTreeV3"
+            libname = "behaviortree_cpp_v3"
+        self.cpp_info.set_property("cmake_file_name", cmake_file_name)
         self.cpp_info.set_property("cmake_target_name", f"BT::{libname}")
+
+        requires = [
+            "zeromq::zeromq",
+            "ncurses::ncurses",
+        ]
+        if self.options.with_coroutines:
+            requires.append("boost::coroutine")
+        if self._with_lexy:
+            requires.append("foonathan-lexy::foonathan-lexy")
+        if self._with_minitrace:
+            requires.append("minitrace::minitrace")
+        if self._with_sqlite3:
+            requires.append("sqlite3::sqlite3")
+        if self._with_tinyxml2:
+            requires.append("tinyxml2::tinyxml2")
 
         postfix = "d" if self.settings.os == "Windows" and self.settings.build_type == "Debug" else ""
         # TODO: back to global scope in conan v2 once cmake_find_package* generators removed
         self.cpp_info.components[libname].libs = [f"{libname}{postfix}"]
-        self.cpp_info.components[libname].requires = [
-            "zeromq::zeromq",
-            "cppzmq::cppzmq",
-            "ncurses::ncurses",
-            "minitrace::minitrace",
-            "tinyxml2::tinyxml2",
-            "foonathan-lexy::foonathan-lexy",
-        ]
-        if Version(self.version) >= "4.1.1":
-            self.cpp_info.components[libname].requires.append("sqlite3::sqlite3")
-        if self.options.with_coroutines:
-            self.cpp_info.components[libname].requires.append("boost::coroutine")
+        self.cpp_info.components[libname].requires = requires
         if self.settings.os in ("Linux", "FreeBSD"):
             self.cpp_info.components[libname].system_libs.append("pthread")
         if Version(self.version) >= "4.0" and \
@@ -201,17 +232,11 @@ class BehaviorTreeCPPConan(ConanFile):
 
         if self.options.with_tools:
             bin_path = os.path.join(self.package_folder, "bin")
-            self.output.info("Appending PATH env var with : {}".format(bin_path))
             self.env_info.PATH.append(bin_path)
 
         # TODO: to remove in conan v2 once cmake_find_package* generators removed
-        if Version(self.version) < "4.0":
-            self.cpp_info.filenames["cmake_find_package"] = "BehaviorTreeV3"
-            self.cpp_info.filenames["cmake_find_package_multi"] = "BehaviorTreeV3"
-        else:
-            self.cpp_info.filenames["cmake_find_package"] = "BehaviorTree"
-            self.cpp_info.filenames["cmake_find_package_multi"] = "BehaviorTree"
-
+        self.cpp_info.filenames["cmake_find_package"] = cmake_file_name
+        self.cpp_info.filenames["cmake_find_package_multi"] = cmake_file_name
         self.cpp_info.names["cmake_find_package"] = "BT"
         self.cpp_info.names["cmake_find_package_multi"] = "BT"
         self.cpp_info.components[libname].names["cmake_find_package"] = libname
